move DamageFromOneToAnotherTargetEffect to common effects
